+++ env.production
@@ -1,0 +1,12 @@
+API_URL=http://example-production.com
+
+
+# public environment key for client side
+NEXT_PUBLIC_FIREBASE_API_KEY=AIzaSyBpQ1K6hW25_2ifd5ucppWnqzhys0-PLsA
+NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN=nextjs-starter-60705.firebaseapp.com
+NEXT_PUBLIC_FIREBASE_DATABASE_URL=https://nextjs-starter-60705.firebaseio.com
+NEXT_PUBLIC_FIREBASE_PROJECT_ID=nextjs-starter-60705
+NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET=nextjs-starter-60705.appspot.com
+NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID=87263525778
+NEXT_PUBLIC_FIREBASE_APP_ID=1:87263525778:web:ce5d5375b2cb1b42e05f4b
+NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID=G-9RFRQQ4QJM
